flask_app/requirements.txt
blinker==1.6.3
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
Flask-Caching==2.1.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.1
